+++ app/Views/admin/unified_booking_management.php
@@ -75,6 +75,9 @@ require_once __DIR__ . '/../partials/header.php';
                             <a href="?controller=admin&action=unifiedBookingManagement" class="btn btn-outline-secondary">
                                 <i class="fas fa-times"></i> Clear
                             </a>
+                            <a href="?controller=admin&action=showOnSiteBookingForm" class="btn btn-success ms-2">
+                                <i class="fas fa-plus"></i> On-Site Book
+                            </a>
                         </div>
                     </form>
                 </div>